sesion25/requirements.txt
django
mercadopago
djangorestframework